--- a/package.json
+++ b/package.json
@@ -108,7 +108,8 @@
     "ts-node": "^10.7.0",
     "tsd": "^0.29.0",
     "typedoc": "^0.23.15",
-    "typescript": "~5.0.4"
+    "typescript": "~5.0.4",
+    "web3": "^4.16.0"
   },
   "packageManager": "yarn@3.8.5",
   "engines": {