--- a/src/index.ts
+++ b/src/index.ts
@@ -36,3 +36,4 @@ export * from './superstruct';
 export * from './time';
 export * from './transaction-types';
 export * from './versions';
+export * from './unitsConversion';